further improve barb rendering; error mark fn takes e,n as coords; border around legend example; determine raster e,n coords from col,row; fix arrow heading convention, drawing argument types
git-svn-id: https://svn.osgeo.org/grass/grass-addons@45758 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.barb/local_proto.h
@@ -22,7 +22,7 @@ double max_magnitude(double *, int);
 
 /* draw.c */
 void draw_barb(double, double, double, double, int, double, int);
-void unknown_(int, int);
+void unknown_(double, double);
 void arrow_mag(double, double, double, double, int);
 void arrow_360(double, double, double, int, double, int);
 void mark_the_spot(double, double);
